requirements.txt
nest-asyncio
numpy
torch
transformers
fastapi
uvicorn
python-dotenv
llama-index
llama-parse
llama-index-vector-stores-faiss
faiss-cpu